refactor(trusty-mpm): extract activity handler submodule + surface task/cwd in SessionSummary
Prep for FIX A/B/C (issue #1787):
- Extract ActivityResponse + get_session_activity into managed_routes/activity.rs
  (mod.rs was at 499 SLOC — 1 under cap — making room for the new fields)
- Add task: Option<String> and cwd: Option<String> to SessionSummary with serde defaults
  so CLI table and info fallback have the task description and working directory
- Populate task/cwd in record_to_summary() and record_to_json()
- Add DecommissionResponse struct with workspace_removed bool and workspace_path_was
  field (FIX C prep); update decommission handler to populate both fields
- Add task/cwd to ManagedSessionSummary client type with #[serde(default)] for
  backward compat with older daemons
